requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
azure-common==1.1.28
azure-core==1.31.0
azure-functions==1.21.3
azure-identity==1.17.1
azure-quantum==2.1.1
azure-storage-blob==12.20.0
azure-synapse-spark==0.7.0
babel==2.16.0
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.1
bleach==6.1.0
blinker==1.9.0
catboost==1.2.7
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dask==2024.10.0
dask-expr==1.1.16
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.7.0
eventlet==0.37.0
executing==2.1.0
fastjsonschema==2.20.0
Flask==3.0.3
Flask-SocketIO==5.4.1
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
gevent==24.11.1
gevent-websocket==0.10.1
graphviz==0.20.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipympl==0.9.4
ipython==8.27.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kafka-python==2.0.2
kiwisolver==1.4.7
kombu==5.4.2
lightgbm==4.5.0
locket==1.0.0
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
patsy==0.5.6
pika==1.3.2
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
portalocker==2.10.1
prometheus_client==0.20.0
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.0.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyodbc==5.1.0
pyparsing==3.1.4
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.10.1
python-json-logger==2.0.7
python-markdown-math==0.8
python-pptx==1.0.2
python-socketio==5.11.4
pytz==2024.2
pywin32==307
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
qsharp==1.8.0
redis==5.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
simple-websocket==1.1.0
six==1.16.0
snapml==1.15.6
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
statsmodels==0.14.4
swifter==1.4.0
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toolz==1.0.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
vine==5.1.0
waitress==3.0.1
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.1
XlsxWriter==3.2.0
yarl==1.17.1
zipp==3.20.2
zope.event==5.0
zope.interface==7.1.1
